requirements.txt
streamlit>=1.28.0
langchain>=0.0.340
langchain-openai>=0.0.2
openai>=1.3.0
pandas>=2.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
pathlib2>=2.3.0
pydantic>=2.0.0